+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 2.0.6 -> 2.0.7:
 * Improved translation subsystem
 * Revamped simcore_init_ui() API function 
+* Improved command line version (progressive output rather than wait to the end)
 
 2.0.5 -> 2.0.6:
 * Initial support for the WepSIM control API